Diffuse light to separate module
